requirements.txt
fastapi
uvicorn
python-dotenv
openai
elevenlabs
llama-index-core
llama-index-readers-file
llama-index-embeddings-openai
llama-index-vector-stores-faiss
faiss-cpu
python-multipart
jinja2
pydantic
llama-index-llms-openai
numpy
tqdm